+++ kotlin/src/main/kotlin/aoc2019/intcode/Computer.kt
@@ -9,11 +9,11 @@ data class Computer(
     private val memory: Memory = Memory(input)
 
     private var last: Memory = memory.copy()
-    fun readResult(index: Int = 0): Int = last[index]
+    fun readResult(index: Int = 0): Long = last[index]
 
     suspend fun run(
-        noun: Int = memory[1],
-        verb: Int = memory[2],
+        noun: Long = memory[1],
+        verb: Long = memory[2],
         io: IO = DefaultIO()
     ) {
         last = memory.copy()
