+++ kotlin/src/main/kotlin/aoc2019/intcode/io/ChannelIO.kt
@@ -3,13 +3,13 @@ package aoc2019.intcode.io
 import kotlinx.coroutines.channels.Channel
 
 class ChannelIO(
-    settings: Array<Int>,
-    private val provider: Channel<Int>,
-    private val sender: Channel<Int>
+    settings: Array<Long>,
+    private val provider: Channel<Long>,
+    private val sender: Channel<Long>
 ): IO(settings) {
-    override suspend fun provide(): Int = provider.receive()
+    override suspend fun provide(): Long = provider.receive()
 
-    override suspend fun write(out: Int) {
+    override suspend fun write(out: Long) {
         sender.send(out)
     }
 }
