+++ kotlin/src/main/kotlin/aoc2019/intcode/io/IO.kt
@@ -1,15 +1,15 @@
 package aoc2019.intcode.io
 
-abstract class IO(private val settings: Array<Int>) {
+abstract class IO(private val settings: Array<Long>) {
     private var pointer = -1
 
-    abstract suspend fun provide(): Int
+    abstract suspend fun provide(): Long
 
-    suspend fun read(): Int {
+    suspend fun read(): Long {
         pointer++
         if (pointer < settings.size) return settings[pointer]
         return provide()
     }
 
-    abstract suspend fun write(out: Int)
+    abstract suspend fun write(out: Long)
 }
